+++ activerecord/lib/arel/nodes/matches.rb
@@ -3,6 +3,8 @@
 module Arel # :nodoc: all
   module Nodes
     class Matches < Binary
+      include FetchAttribute
+
       attr_reader :escape
       attr_accessor :case_sensitive
 
@@ -11,8 +13,16 @@ module Arel # :nodoc: all
         @escape = escape && Nodes.build_quoted(escape)
         @case_sensitive = case_sensitive
       end
+
+      def invert
+        Arel::Nodes::DoesNotMatch.new(left, right, escape, case_sensitive)
+      end
     end
 
-    class DoesNotMatch < Matches; end
+    class DoesNotMatch < Matches
+      def invert
+        Arel::Nodes::Matches.new(left, right, escape, case_sensitive)
+      end
+    end
   end
 end